+++ tests/fullSentence.test.js
@@ -26,3 +26,15 @@ test('get full sentence:', function(t) {
 
   t.end()
 })
+
+test('get multiple-copies of one sentence:', function(t) {
+  let doc = nlp('John Smith was cool. I am missing. Cindy Lauper and Carl Sagan here. I am also missing.')
+  let m = doc.match('#Person+')
+  let matches = m.fullSentences()
+  let arr = matches.out('array')
+  t.equal(arr.length, 3, 'two sentences into three results')
+  t.equal(arr[0], 'John Smith was cool.', 'one person sentence #1')
+  t.equal(arr[1], 'Cindy Lauper and Carl Sagan here.', 'two person sentence #1')
+  t.equal(arr[2], 'Cindy Lauper and Carl Sagan here.', 'two person sentence #2')
+  t.end()
+})