+++ changelog.md
@@ -24,6 +24,7 @@ While all _Major_ releases should be reviewed, our only two _large_ releases are
 - support lexicon object input to tokenize - `.tokenize({word:'tag'})`
 - clean-up of scripts and tooling
 - improved typescript types
+- add `.sentence()` method to retrieve full sentence from a match
 -->
 
 ##### 12.4.0